requirements.txt
sounddevice
numpy
scipy
tk
assemblyai
google-generativeai
python-dotenv